requirements.txt
python-telegram-bot==21.2
requests==2.32.3